🚚 Move constants to @/utils/
