+++ src/components/core/lume-point/lume-point.vue
@@ -20,7 +20,7 @@ import { computed, onMounted, PropType, ref } from 'vue';
 
 import { Scale } from '@/composables/scales';
 
-import { Colors } from '@/constants';
+import { Colors } from '@/utils/constants';
 import { getScaleStep, isBandScale } from '@/utils/helpers';
 import { svgCheck } from '@/utils/svg-check';
