+++ src/composables/props.ts
@@ -4,7 +4,7 @@ import { DataValidator, withBase } from './base';
 import { ChartOptions, Options, withOptions } from './options';
 import { withScales } from './scales';
 
-import { Orientation, ORIENTATIONS } from '@/constants';
+import { Orientation, ORIENTATIONS } from '@/utils/constants';
 
 function orientationValidator(orientation: string): boolean {
   return Object.values(ORIENTATIONS).includes(orientation as Orientation);
