+++ src/next/box-plot/defaults.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_MARGINS } from '@/constants';
+import { DEFAULT_MARGINS } from '@/utils/constants';
 
 export const options = {
   xAxisOptions: {},
