+++ src/playground/line-bar-chart.stories.ts
@@ -1,5 +1,5 @@
 import LineBarChart from './line-bar-chart.vue';
-import { DEFAULT_MARGINS } from '@/constants';
+import { DEFAULT_MARGINS } from '@/utils/constants';
 
 export default {
   title: 'Playground/Line+bar chart',
