+++ test/unit/mock-data.ts
@@ -2,7 +2,7 @@ import { Data, DatasetValueObject } from '@/types/dataset';
 import { Scale } from '@/composables/scales';
 import { scaleBand, scaleLinear } from 'd3';
 import { ContainerSize } from '@/types/size';
-import { LegacyColors } from '@/constants';
+import { LegacyColors } from '@/utils/constants';
 
 const width = 640;
 const height = 480;
